NodeLinks
Here is a very basic NodeLinks file for constructing nodes for a Linked List. Over the weekend I constructed a few linked lists from scratch, via youtube tutorials, but that seems a bit of an overkill for our assignment. And, I believe all the built in methods that Java provides will be better anyway. So, we'll have to call the linked list method throughout the code as needed, but this gives us a way to create new nodes at the very least. I look forward to a conversation on Tuesday as I'm still a bit unsure how everything will fit together. Specifically, we discussed the possibility that the Linked List would be a Linked List of Linked Lists (fun to say out-loud), buy after looking over things again, it appears it will only be a linked list of nodes. I may be looking at this wrong, but will have more time tonight for examination.